requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
gunicorn==22.0.0
